Wait for complete transmission of PRE
.. by looking for the related state machine to enter txstall

h/t @tannewt for doing a protocol trace and noticing that there was
no LS packet transmitted after the PRE. I don't know what those words
mean but it led to me conjecturing a fix.

I also moved the clear of the IRQ complete flag above the start of the
transfer, just in case the transfer could complete before the flag was
cleared. But, I don't think this was a cause of the problem.

it may be worth looking at the other sites where IRQ_TX_EOP_MASK is used
to see whether similar treatment is necessary.

With this, a LS keyboard & LS gamepad will enumerate on my FruitJam
at 240MHz (via tools menu) & at 268MHz with HDMI display (dvhstx hard-coded
overclock)